+++ src/main/java/com/green/campingsmore/order/payment/model/PaymentCompleteDto.java
@@ -7,5 +7,5 @@ public class PaymentCompleteDto {
     private String address;
     private String addressDetail ;
     private String shippingMemo;
-    private Long totalPrice;
+    private Integer totalPrice;
 }
